+++ android/app/src/main/java/com/hebarcode/reader/MainApplication.kt
@@ -4,7 +4,6 @@ import android.app.Application
 import com.facebook.react.PackageList
 import com.facebook.react.ReactApplication
 import com.facebook.react.ReactHost
-import com.facebook.react.ReactPackage
 import com.facebook.react.ReactNativeApplicationEntryPoint.loadReactNative
 import com.facebook.react.defaults.DefaultReactHost.getDefaultReactHost
 
@@ -15,7 +14,7 @@ class MainApplication : Application(), ReactApplication {
       context = applicationContext,
       packageList =
         PackageList(this).packages.apply {
-          add(HebarcodeScannerPackage() as ReactPackage)
+          add(HebarcodeScannerPackage())
         },
     )
   }